Add code executing

+++ OperatingSystem/ResourceManagement/ResourceParts/InterruptData.cs
@@ -3,5 +3,5 @@ namespace OperatingSystem.ResourceManagement.ResourceParts;
 public class InterruptData : ResourcePart
 {
     public Guid JobGovernorGuid { get; set; }
-    public int InterruptCode { get; set; }
+    public byte InterruptCode { get; set; }
 }
