+++ OperatingSystem/ResourceManagement/ResourceParts/JobGovernorInterruptData.cs
@@ -2,5 +2,5 @@ namespace OperatingSystem.ResourceManagement.ResourceParts;
 
 public class JobGovernorInterruptData : ResourcePart
 {
-    public int InterruptCode { get; set; }
+    public byte InterruptCode { get; set; }
 }
